Extract RoledControllerPolicy
