+++ spec/known_dirs.txt
@@ -13,3 +13,4 @@ spec/services
 spec/workers
 spec/counters
 spec/routes
+spec/policies
